% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{abname_molis}
\alias{abname_molis}
\title{Name of antibiotic}
\source{
\code{\link[AMR]{antibiotics}}
}
\usage{
abname_molis(abcode, textbetween = " + ", tolower = FALSE, ...)
}
\arguments{
\item{abcode}{antibiotic code or name.}

\item{textbetween}{text between two or more antibiotics.}

\item{tolower}{result in lower case.}

\item{...}{arguments for \code{\link[=abname_molis]{abname_molis()}}}
}
\description{
Transforms a code to a antibiotic name or ATC-code.
}
\examples{
abname_molis("amcl")
# "Amoxicilline/clavulaanzuur"

abname_molis("amcl+gent")
# "Amoxicilline/clavulaanzuur + gentamicine"

abname_molis(c("amox", "amcl"))
# "Amoxicilline" "F"
}
\keyword{ab}
\keyword{antibiotics}
