% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{agb_property}
\alias{agb_property}
\title{Retrieve AGD Property}
\usage{
agb_property(agb_codes, property = NULL)
}
\arguments{
\item{agb_codes}{AGB codes}

\item{property}{property of the AGB code. Will return a \link{data.frame} if left blank.}
}
\description{
Download properties from the \href{https://www.vektis.nl/agb-register}{AGB-register}.
}
\examples{
agb_property(03033048)
agb_property(c(03033048, 01102504))

agb_property(03033048, property = "last_name")
agb_property(c(03033048, 01102504), "specialty")
}
