% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{as.UTC}
\alias{as.UTC}
\alias{as.UTC.data.frame}
\alias{as.UTC.POSIXct}
\alias{as.UTC.default}
\title{Force Time as UTC}
\usage{
as.UTC(x, ...)

\method{as.UTC}{data.frame}(x, ...)

\method{as.UTC}{POSIXct}(x, ...)

\method{as.UTC}{default}(x, ...)
}
\arguments{
\item{x}{a vector of datetime values}

\item{...}{not used at the moment}
}
\description{
Force Time as UTC
}
