% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{as_excel}
\alias{as_excel}
\alias{save_excel}
\title{Create Excel Document}
\usage{
as_excel(
  ...,
  sheet_names = NULL,
  autofilter = TRUE,
  autowidth = TRUE,
  rows_zebra = TRUE,
  cols_zebra = FALSE,
  save = identical(deparse(substitute(...)), "."),
  file = NULL,
  freeze_top_row = TRUE,
  overwrite = FALSE
)

save_excel(xl, file = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{...}{data objects, use named items for multiple tabs (see \emph{Examples}) zie Examples}

\item{sheet_names}{sheet names}

\item{autofilter}{create autofilter on columns in first row}

\item{autowidth}{automatically adjust columns widths}

\item{rows_zebra}{create banded rows}

\item{cols_zebra}{create banded columns}

\item{save}{save the file directly, defaults to \code{TRUE} if used after a pipe (\verb{\%>\%})}

\item{file}{file location to save Excel document to, defaults to a random filename}

\item{freeze_top_row}{freeze the first row of the sheet}

\item{overwrite}{overwrite existing file}

\item{xl}{Excel object, as created with the \code{openxlsx} package}
}
\description{
This function relies on the \code{openxlsx} package for creating Excel documents.
}
\examples{
# does not save immediately, but returns the workbook object
xl <- as_excel("this is a sheet" = mtcars,
               "another sheet" = anscombe)
xl


# save with save_excel() or export.excel()
export.excel(xl)
         
\dontrun{
# saves immediately
if (require("dplyr")) {
  mtcars \%>\%
    as_excel()
}
}
}
