% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{auto_transform}
\alias{auto_transform}
\title{Automatically Transform Data Set}
\usage{
auto_transform(
  x,
  datenames = "en",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ".",
  big.mark = "",
  timezone = "",
  na = c("", "NULL", "NA", "<NA>"),
  snake_case = FALSE,
  ...
)
}
\arguments{
\item{x}{a \link{data.frame}}

\item{datenames}{language of the date names, such as weekdays and months}

\item{dateformat}{expected date format, will be coerced with \code{\link[cleaner:format_datetime]{format_datetime()}}}

\item{timeformat}{expected time format, will be coerced with \code{\link[cleaner:format_datetime]{format_datetime()}}}

\item{decimal.mark}{separator for decimal numbers}

\item{big.mark}{separator for thousands}

\item{timezone}{expected time zone}

\item{na}{values to interpret as \code{NA}}

\item{snake_case}{apply \href{https://en.wikipedia.org/wiki/Snake_case}{snake case} to the column names}

\item{...}{not used as the time, allows for future extension}
}
\description{
This function transforms a \link{data.frame} by guessing the right data classes and applying them, using \code{\link[readr:parse_guess]{readr::parse_guess()}} and \code{cleaner} functions such as \code{\link[cleaner:clean]{cleaner::clean_Date()}}.
}
