% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_prevention.R
\name{.case_when}
\alias{.case_when}
\title{Case When}
\source{
Code copied from \code{\link[poorman:case_when]{poorman::case_when()}} under same license.
}
\usage{
.case_when(...)
}
\arguments{
\item{...}{a sequence of two-sided formulas}
}
\description{
Inspired by \code{\link[dplyr:case_when]{dplyr::case_when()}}, but allows different lengths of outcomes (just like \code{\link[=ifelse]{ifelse()}}).
}
\examples{
x <- 1:50
.case_when(
  x \%\% 35 == 0 ~ "fizz buzz",
  x \%\% 5 == 0 ~ "fizz",
  x \%\% 7 == 0 ~ "buzz",
  TRUE ~ as.character(x)
)
}
