% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs.R
\name{cbs_topics}
\alias{cbs_topics}
\alias{cbs_search}
\alias{cbs_download}
\alias{cbs_moreinfo}
\title{Download CBS-data}
\usage{
cbs_topics()

cbs_search(topic, max_print = 25)

cbs_download(identifier, clean_cols = TRUE)

cbs_moreinfo(identifier)
}
\arguments{
\item{topic}{topics to search for}

\item{max_print}{maximum number of subjects to print}

\item{identifier}{tracking number (1 to \code{max_print}) in \code{\link[=cbs_search]{cbs_search()}}}

\item{clean_cols}{Clean column names.}
}
\description{
Download data from \href{https://www.cbs.nl/nl-nl/onze-diensten/open-data/statline-als-open-data}{CBS Open data Statline}.
}
\details{
\code{\link[=cbs_topics]{cbs_topics()}} retrieves all topics.

\code{\link[=cbs_search]{cbs_search()}} searches for a specific subject.

\code{\link[=cbs_download]{cbs_download()}} downloads tables. Input has to be a CBS Identifier (printed in red in \code{\link[=cbs_search]{cbs_search()}}), or a tracking number of \code{\link[=cbs_search]{cbs_search()}}, or the result of \code{\link[=cbs_search]{cbs_search()}}.

\code{\link[=cbs_moreinfo]{cbs_moreinfo()}} gives a detailed explanation for the table. Input can also be a dataset downloaded with \code{\link[=cbs_download]{cbs_download()}}.
}
\examples{
cbs_search("Inwoners")

x <- cbs_download(2) # 2nd hit of cbs_search()
str(x)

cbs_moreinfo(x)
cbs_moreinfo(2)
}
