% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certetoolbox-package}
\alias{certetoolbox}
\alias{certetoolbox-package}
\title{certetoolbox: A Certe R Package for Miscellaneous Functions}
\description{
A Certe R Package for miscellaneous functions that do not fit a dedicated package. This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certetoolbox}
  \item \url{https://github.com/certe-medical-epidemiology/certetoolbox}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing \email{e.hassing@certe.nl}
  \item Erwin Dijkstra \email{e.dijkstra@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
