% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_mo}
\alias{check_mo}
\title{Check micro-organism from early warning mail}
\usage{
check_mo(ordernr_testcode_mtrlcode_stam, view = TRUE, info = FALSE, ...)
}
\arguments{
\item{ordernr_testcode_mtrlcode_stam}{strain ID}

\item{view}{direct view in RStudio viewer.}

\item{info}{show query information.}

\item{...}{arguments passed on to \code{\link{certedb_query}}.}
}
\description{
Use this function to check results and resistance pattern of a strain using the ID provided in the early warning mail.
}
