% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal.R, R/universal_extra.R
\name{concat}
\alias{concat}
\alias{collapse}
\title{Paste Character Vectors Together}
\usage{
concat(..., sep = "")

collapse(x, sep = "")
}
\arguments{
\item{..., x}{element(s) to be pasted together, can also be vectors}

\item{sep}{separator character, will also be used for collapsing}
}
\description{
The \code{\link[=concat]{concat()}} function is at default identical to \code{paste(c(...), sep = "", collapse = "")}.

The \code{\link[=collapse]{collapse()}} function is at default identical to \code{paste(x, sep = "", collapse = "")}.
}
\examples{
concat("a", "b", "c")

concat(c("a", "b"), "c")
collapse(c("a", "b"), "c")

concat(letters[1:5], "-")
collapse(letters[1:5], "-")
}
