% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{create_confusionmatrix}
\alias{create_confusionmatrix}
\title{Create a Confusion Matrix}
\usage{
create_confusionmatrix(
  df,
  identifier,
  compare,
  outcome,
  positive = "^pos.*",
  negative = "^neg.*",
  ...,
  ignore_case = TRUE
)
}
\arguments{
\item{df}{a \link{data.frame}}

\item{identifier}{a column name to use as identifier, such as a patient ID or an order ID}

\item{compare}{a column name for the two axes of the table: the labels between the outcomes must be compared}

\item{outcome}{a column name containing the outcome values to compare}

\item{positive}{a \link{regex} to match the values in \code{outcome} that must be considered as the Positive class, use \code{FALSE} to not use a Positive class}

\item{negative}{a \link{regex} to match the values in \code{outcome} that must be considered as the Negative class, use \code{FALSE} to not use a Negative class}

\item{...}{manual \link{regex}es for classes if not using \code{positive} and \code{negative}, such as \verb{Class1 = "c1", Class2 = "c2", Class3 = "c3"}}

\item{ignore_case}{a \link{logical} to indicate whether the case in the values of \code{positive}, \code{negative} and \code{...} must be ignored}
}
\description{
Transform a data set into a \emph{n} x \emph{m} table, e.g. to be used in \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}}.
}
\examples{
df <- data.frame(
  order_nr = sort(rep(LETTERS[1:20], 2)),
  test_type = rep(c("Culture", "PCR"), 20),
  result = sample(c("pos", "neg"),
                  size = 40,
                  replace = TRUE,
                  prob = c(0.3, 0.9))
)
head(df)

out <- df |> create_confusionmatrix(order_nr, test_type, result)
out


df$result <- gsub("pos", "#p", df$result)
df$result <- gsub("neg", "#n", df$result)
head(df)
# gives a warning that pattern matching failed:
df |> create_confusionmatrix(order_nr, test_type, result)

# define the pattern yourself in such case:
df |> create_confusionmatrix(order_nr, test_type, result,
                             positive = "#p",
                             negative = "#n")
                             
                             
# defining classes manually, can be more than 2:
df |> create_confusionmatrix(order_nr, test_type, result,
                             ClassA = "#p", Hello = "#n")
                             
if ("caret" \%in\% rownames(utils::installed.packages())) {
  # confusionMatrix() is also re-exported by certestats
  caret::confusionMatrix(out)
}
}
