% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{crosstab}
\alias{crosstab}
\title{Create contingency table of \code{data.frame}}
\usage{
crosstab(
  data,
  column1,
  condition1,
  column2,
  condition2,
  expected = FALSE,
  totals = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble} containing data, where columns \code{column1} and \code{column2} occur.}

\item{column1}{Column with values.}

\item{condition1}{Condition to seperate \code{column1} from "Rest".}

\item{column2}{Column with values.}

\item{condition2}{Condition to seperate \code{column2} from "Rest".}

\item{expected}{Default is \code{FALSE}. Return expected instead of observed (\code{E_i = N / n}, where discrete uniforme distribution is to be expected).}

\item{totals}{Default is \code{FALSE}. Adds row- and columntotals.}
}
\value{
\code{matrix}
}
\description{
Creates a contingency table. Output is a \code{\link{matrix}}.
}
\keyword{chi}
\keyword{contingency}
\keyword{squared}
\keyword{table}
