% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{export}
\alias{export}
\alias{export.R}
\alias{export.excel}
\alias{export.xlsx}
\alias{export.csv}
\alias{export.csv2}
\alias{export.tsv}
\alias{export.spss}
\alias{export.pdf}
\alias{export.png}
\alias{export.clipboard}
\title{Export Data and Plots}
\usage{
export(
  object,
  fn,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  ...
)

export.R(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  ...
)

export.excel(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  sheet_names = NULL,
  autofilter = TRUE,
  rows_zebra = TRUE,
  cols_zebra = FALSE,
  ...
)

export.xlsx(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  sheet_names = NULL,
  autofilter = TRUE,
  rows_zebra = TRUE,
  cols_zebra = FALSE,
  ...
)

export.csv(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  na = "",
  ...
)

export.csv2(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  na = "",
  ...
)

export.tsv(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  na = "",
  ...
)

export.spss(
  object,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  export_qry = TRUE,
  ...
)

export.pdf(
  plot,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  size = "A5",
  portrait = FALSE,
  ...
)

export.png(
  plot,
  filename = NULL,
  card_number = project_get_current_id(ask = FALSE),
  width = 1000,
  height = 800,
  text.factor = 1.2,
  ...
)

export.clipboard(
  object,
  sep = "\\t",
  na = "",
  header = TRUE,
  quote = FALSE,
  decimal.mark = ",",
  ...
)
}
\arguments{
\item{object, plot}{the \R object to export}

\item{fn}{a manual export function, such as \code{haven::write_sas} to write SAS files}

\item{filename}{the full path of the exported file}

\item{card_number}{a Trello card number}

\item{...}{arguments passed on to methods}

\item{export_qry}{export the underlying SQL query as well}

\item{sheet_names}{sheet names}

\item{autofilter}{create autofilter on columns in first row}

\item{rows_zebra}{create banded rows}

\item{cols_zebra}{create banded columns}

\item{na}{replacement character for empty values (default: \code{""})}

\item{size}{paper size, defaults to A5. Can be A0 to A7.}

\item{portrait}{portrait mode, defaults to \code{FALSE} (i.e., landscape mode)}

\item{width}{required width of the PNG file in pixels}

\item{height}{required height of the PNG file in pixels}

\item{text.factor}{text factor for the exported plot. Defaults to \code{1.2}, which loosely equals a PDF file in A5 format when it comes to text sizes.}

\item{sep}{(for \code{\link[=export.clipboard]{export.clipboard()}}) separator for values in a row (default: tab)}

\item{header}{(for \code{\link[=export.clipboard]{export.clipboard()}}) use column names as header (default: \code{TRUE})}

\item{quote}{(for \code{\link[=export.clipboard]{export.clipboard()}}) use quotation marks (default: \code{FALSE})}

\item{decimal.mark}{(for \code{\link[=export.clipboard]{export.clipboard()}}) character to use for decimal numbers}
}
\description{
These functions can be used to export data and plots. They invisibly return the object itself again, allowing for usage in pipes. The functions work closely with the \code{certeprojects} package and support Trello card numbers.
}
\details{
The \code{\link[=export]{export()}} function can export to any file format, as long as the export function is passed on to the \code{fn} argument. This function has to have the object as first argument and the future file location as second argument.

RDS files as created by \code{\link[=export.R]{export.R()}} are compatible with R3 and R4.

The \code{\link[=export.excel]{export.excel()}} uses \code{\link[=as_excel]{as_excel()}} internally.

For \code{\link[=export.csv]{export.csv()}}, \code{\link[=export.csv2]{export.csv2()}} and \code{\link[=export.tsv]{export.tsv()}}, files will be saved in UTF-8 encoding and values \code{NA} will be exported as \code{""}. Like other \verb{*.csv} and \verb{*.csv2} functions, csv is comma (\verb{,}) separated and csv2 is semicolon (\verb{;}) separated.

Exporting as SPSS files with \code{\link[=export.spss]{export.spss()}} requires the \code{haven} package to be installed.

If the filename is left blank in \code{\link[=export.pdf]{export.pdf()}} or \code{\link[=export.png]{export.png()}}, the title of \code{plot} will be used if available (and a timestamp otherwise).

The \code{\link[=export.clipboard]{export.clipboard()}} function requires the \code{clipr} package to be installed. The function allows any object (also other than \link{data.frame}s) and is only limited to the available amount of RAM memory.
}
\examples{
\dontrun{
library(dplyr)

# export two files: 'whole_file.rds' and 'first_ten_rows.xlsx'
starwars \%>\%
  export.R("whole_file") \%>\%
  slice(1:10) \%>\%
  export.excel("first_ten_rows")
}
}
