% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{generate_hash}
\alias{generate_hash}
\title{Quick SHA Hash}
\usage{
generate_hash(x, key = NULL, bits = 256)
}
\arguments{
\item{x}{input, will be converted to character}

\item{key}{key or salt}

\item{bits}{number of bits, often one of: 224, 256, 384, 512}
}
\description{
Generates hashes with or without an extra key/salt.
}
\details{
This function uses the \code{\link[openssl:hash]{openssl::sha2()}} function, but always returns a \link{character} vector.
}
\examples{
generate_hash("a")
generate_hash(c("a", "b", "c"))

generate_hash("a", "secret")
generate_hash(c("a", "b", "c"), "secret")
}
