% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{generate_identifier}
\alias{generate_identifier}
\title{Create Random Identifier}
\usage{
generate_identifier(id_length = 6, n = 1, chars = c(0:9, letters[1:6]))
}
\arguments{
\item{id_length}{character length of ID}

\item{n}{number of IDs to generate}

\item{chars}{characters to use for generation, defaults to hexadecimal characters (0-9 and a-f)}
}
\description{
This function creates unique identifier (IDs) using \code{\link[=sample]{sample()}}.
}
\examples{
generate_identifier(8)
generate_identifier(6, 3)
}
