% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{hospital_name}
\alias{hospital_name}
\title{Hospitalname}
\usage{
hospital_name(x, format = "{naamkort}, {plaats}")
}
\arguments{
\item{x}{text to be transformed}

\item{format}{default is \code{"{naamkort}, {plaats}"}. Attributes like \code{x} to be returned in '\code{glue}'-format (in curly brackets).}
}
\description{
Hospitalname and/or location, with support for all hospitals in Northern Netherlands, including Meppel, Hardenberg and Zwolle.
}
\examples{
hospital_name(c("MCL", "MCL", "Martini"))
hospital_name(c("Antonius", "WZA", "Martini"), format = "{naam} te {plaats}")

# special case for GGD
hospital_name(c("Martini", "GGD Groningen", "GGD Drenthe"), format = "{naam}")
hospital_name(c("Martini", "GGD Groningen", "GGD Drenthe"), format = "{naamkort}")
hospital_name("ggd friesland", "{naam}")
}
