% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{json_data_structure}
\alias{json_data_structure}
\title{JSON Dataset Structure}
\usage{
json_data_structure(dataset, dataset_name = NULL, output_file = NULL)
}
\arguments{
\item{dataset}{\link{data.frame}}

\item{dataset_name}{name of \code{dataset}, can be left blank to auto-guess}

\item{output_file}{path to the output JSON file.}
}
\value{
Invisibly returns the named list written to JSON.
}
\description{
Generates a JSON representation of a dataset’s structure, capturing unique values for each column unless the column appears to be an identifier (i.e., has as many unique values as rows) or is a date/time column. All columns are included in the output.
}
