% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{mo_certe}
\alias{mo_certe}
\title{Certe bacteria code}
\usage{
mo_certe(x, ...)
}
\arguments{
\item{x, ...}{See \code{\link[AMR]{mo_property}}.}
}
\description{
Input is evaluated with \code{\link[AMR]{as.mo}} and compared to the data set \code{certetools:::bacteriecode} to retreive a Certe bacteria code.
}
\details{
Reference created with:
\preformatted{
bacterien <- certedb_query("select * from temporary_certemm_bacterienlijst")

bacterien <- bacterien \%>\%
  transmute(bacteriecode,
            bacteriecode_oud,
            systeemcode,
            fullname = gsub(" species$", "", volledigenaam) \%>\%
              gsub("[)(]", "", .)) \%>\%
  arrange(bacteriecode)

usethis::use_data(bacterien, internal = TRUE, overwrite = TRUE)
}
}
