% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{mo_glims}
\alias{mo_glims}
\title{Microorganisms Code from GLIMS10}
\usage{
mo_glims(
  x,
  language = AMR::get_AMR_locale(),
  keep_synonyms = getOption("AMR_keep_synonyms", FALSE),
  ...
)
}
\arguments{
\item{x}{Any \link{character} (vector) that can be coerced to a valid microorganism code with \code{\link[AMR:as.mo]{as.mo()}}. Can be left blank for auto-guessing the column containing microorganism codes if used in a data set, see \emph{Examples}.}

\item{language}{Language to translate text like "no growth", which defaults to the system language (see \code{\link[AMR:get_AMR_locale]{get_AMR_locale()}}).}

\item{keep_synonyms}{A \link{logical} to indicate if outdated, previously valid taxonomic names must be preserved and not be corrected to currently accepted names. Do note that the term "synonym" is in this case jargon from the field of microbial taxonomy - it is not in place to denote that e.g. "Streptococcus Group A" is a synonym of \emph{S. pyogenes}. Though this is practically the case, taxonomically it is not as "Streptococcus Group A" is not even a valid taxonomic name.

The default is \code{FALSE}, which will return a note if outdated taxonomic names were processed. The default can be set with the package option \code{\link[AMR:AMR-options]{AMR_keep_synonyms}}, i.e. \code{options(AMR_keep_synonyms = TRUE)} or \code{options(AMR_keep_synonyms = FALSE)}.}

\item{...}{Other arguments passed on to \code{\link[AMR:as.mo]{as.mo()}}, such as 'minimum_matching_score', 'ignore_pattern', and 'remove_from_input'.}
}
\description{
This function is analogous to all \verb{mo_*} functions of the AMR package, see \code{\link[AMR:mo_property]{AMR::mo_property()}}.
}
\examples{
mo_glims("E. coli")

library(dplyr, warn.conflicts = FALSE)
data.frame(mo = c("ESCCOL", "Staph aureus")) |>
  mutate(glims = mo_glims())
  
# even works for non-existing entries in AMR package
mo_glims("Streptococcus mitis/oralis")

if (require(AMR, warn.conflicts = FALSE)) {
  as.mo("Streptococcus mitis/oralis")
}
if (require("AMR")) {
  mo_genus("Streptococcus mitis/oralis")
}
if (require("AMR")) {
  mo_gramstain("Streptococcus mitis/oralis")
}
}
