% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_extra.R
\name{notin}
\alias{notin}
\alias{\%notin\%}
\title{Negate \code{\%in\%}}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\description{
Default in R >= 4.6.0.
}
