% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{p_symbol}
\alias{p_symbol}
\title{P-symbol format as asterisk}
\usage{
p_symbol(p, emptychar = " ")
}
\arguments{
\item{p}{numeric value between 0 and 1}

\item{emptychar}{sign to be displayed for 0.1 < p < 1.0}
}
\description{
P-symbol format as asterisk
}
