% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_httr.R
\name{post_data}
\alias{post_data}
\alias{create_json_body}
\title{POST Data (Encrypted)}
\usage{
post_data(
  object,
  url,
  authorization_header = NULL,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)

create_json_body(
  object,
  compress = TRUE,
  encrypt = TRUE,
  key = read_secret("tools.encryption_password")
)
}
\arguments{
\item{object}{Any object of any size, preferably a data set}

\item{url}{A character string specifying the target URL for the HTTP POST request. Must include the full scheme (e.g., \code{"https://"} or \code{"http://"}), hostname, and path.}

\item{authorization_header}{A character string specifying the value of the \code{Authorization} header to include in the POST request, e.g. \code{"Bearer <token>"}. Use \code{NULL} to omit the header.}

\item{compress}{Should the object be compressed/decompressed? At least allowed: \code{"gzip"} (or \code{TRUE}), \code{"bzip2"}, \code{"xz"}, see \code{\link[base:memCompress]{base::memCompress()}}. Use \code{FALSE} to not compress/decompress.}

\item{encrypt}{Should the object be encrypted/decrypted? This applies AES-GCM via \code{\link[openssl:aes_cbc]{openssl::aes_gcm_encrypt()}}, providing authenticated encryption. This guarantees both confidentiality and integrity: the file cannot be read without the correct \code{key}, and any tampering will be detected automatically during decryption. The initialization vector (iv) will be a length-12 random \link{raw} vector.}

\item{key}{A character to be used as the encryption key. Internally, this is converted using \code{\link[openssl:hash]{openssl::sha256()}} to ensure a raw high-entropy key of length \code{32}, suitable for AES-GCM encryption. The default is the \link[=Sys.getenv]{system environment variable}: \code{mmbi_epi_encryption_key}.}
}
\description{
Securely transfer any R object with full attribute preservation and cross-platform JSON compatibility. This uses the same approach as the \href{https://github.com/umcg-mmbi-epidemiology/mmbi.epi/blob/1ed61034ad0c68fd0e78843aa1ff253d141ce6d8/R/serialise-encrypt-post.R}{\code{mmbi.epi} package of the UMCG}.
}
\examples{
\dontrun{

post_data(iris,
          url = "https://some-server:8000/post",
          compress = TRUE,
          encrypt = TRUE)
}

# MANUAL WAY -----------------------------------------------------------

# use create_json_body() to make a(n encrypted) JSON of an object

iris_json <- iris |> create_json_body(compress = TRUE, encrypt = TRUE)

# curl -X POST https://some-server:8000/post
#      -H "Content-Type: application/json"
#      -d '...'

# replace the "..." with the outcome of create_json_body():
iris_json
}
