% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_check.R
\name{privacy_check}
\alias{privacy_check}
\title{Check Privacy of Plain Files}
\usage{
privacy_check(
  path = getwd(),
  vault = paste0(path, "/vault"),
  suspicion = "(postcode|bsn|geboortedatum)"
)
}
\arguments{
\item{path}{path to check}

\item{vault}{path to vault}
}
\description{
Checks if files contain privacy sensitive data and moves them to a 'vault folder' if this is the case.
}
