% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{today}
\alias{now}
\alias{dec_mark}
\alias{big_mark}
\alias{project_get_current_id}
\alias{extract}
\alias{extract2}
\alias{add}
\alias{subtract}
\alias{multiply_by}
\alias{raise_to_power}
\alias{divide_by}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{certeprojects}{\code{\link[certeprojects:project_properties]{project_get_current_id}}}

  \item{certestyle}{\code{\link[certestyle:dec_mark]{big_mark}}, \code{\link[certestyle]{dec_mark}}}

  \item{lubridate}{\code{\link[lubridate]{now}}, \code{\link[lubridate:now]{today}}}

  \item{magrittr}{\code{\link[magrittr:aliases]{add}}, \code{\link[magrittr:aliases]{divide_by}}, \code{\link[magrittr:aliases]{extract}}, \code{\link[magrittr:aliases]{extract2}}, \code{\link[magrittr:aliases]{multiply_by}}, \code{\link[magrittr:aliases]{raise_to_power}}, \code{\link[magrittr:aliases]{subtract}}}
}}

