% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{ref_dir}
\alias{ref_dir}
\title{Return Reference Directory}
\usage{
ref_dir(sub = "")
}
\arguments{
\item{sub}{relative subfolder or file}
}
\description{
Returns the relative reference directory for non-projects.
}
\details{
This function returns the absolute path using \code{\link[tools:fileutils]{tools::file_path_as_absolute()}}.
}
