% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remember}
\alias{remember}
\alias{recall}
\title{Temporarily save value in Global Environment}
\usage{
remember(.data, ..., envir = globalenv())

recall(x = NULL, delete = TRUE, envir = globalenv())
}
\arguments{
\item{.data}{table to be passes through unchanged.}

\item{...}{value(s) to be remembered.}

\item{envir}{default is Global Environment. The \link{environment} where values are saved.}

\item{x}{value to be recalled.}

\item{delete}{delete value after.}
}
\description{
Can be used in dplyr-syntax to remember for later use. Values are temporarily saved in the Global Environment.
}
\details{
values can be saved with \strong{\code{remember()}} and recalled (and deleted) with \strong{\code{recall()}}.
}
\examples{
\dontrun{
 tbl \%>\%
   filter(...) \%>\%
   remember(rows = nrow(.)) \%>\%
   group_by(...) \%>\%
   summarise(...) \%>\%
   plot2(title = "Test",
         subtitle = paste("n =", recall(rows)))
}
}
