% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rsi_table}
\alias{rsi_table}
\title{Susceptibility table between hospitals}
\usage{
rsi_table(
  ab_list,
  hospitalname,
  df_all,
  df_thishospital = NULL,
  df_otherhospitals = NULL
)
}
\arguments{
\item{ab_list}{list of antibiotics. See \code{\link[AMR]{antibiotics}}.}

\item{hospitalname}{name of the hospital to be compared to other hospitals.}

\item{df_all}{\code{data.frame} with all data.}

\item{df_thishospital}{\code{data.frame} with all data of the to be tested hospital.}

\item{df_otherhospitals}{\code{data.frame} with all data of the other hospitals.}
}
\description{
Creates a susceptibility comparison table between hospitals. Runs a G-test at >1000 observations or an Exact-test when less.
}
\examples{
\dontrun{
rsi_table(ab_list, 'MZ', df)

septic_patients \%>\%
  mutate(zkhgroep = hospital_id) \%>\%
  rsi_table(ab_list = c("amox", "amcl"),
            hospitalname = "A",
            df_all = .)
}
}
\seealso{
\code{\link{g.test}} runs at > 1000 observations; \code{\link{exact.test}} runs at <= 1000 observations
}
