% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{tbl_gtsummary}
\alias{tbl_gtsummary}
\title{Summarise Table as \code{gtsummary}}
\usage{
tbl_gtsummary(
  x,
  by = NULL,
  label = NULL,
  digits = 1,
  ...,
  language = "nl",
  column1_name = "Eigenschap",
  add_n = FALSE,
  add_p = FALSE,
  add_ci = FALSE,
  add_overall = FALSE,
  decimal.mark = dec_mark(),
  big.mark = big_mark()
)
}
\arguments{
\item{x}{a \link{data.frame}}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column from \code{data}. Summary statistics will be stratified by this variable.
Default is \code{NULL}.}

\item{label}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Used to override default labels in summary table, e.g. \code{list(age = "Age, years")}.
The default for each variable is the column label attribute, \code{attr(., 'label')}.
If no label has been set, the column name is used.}

\item{digits}{(\code{\link[gtsummary:syntax]{formula-list-selector}})\cr
Specifies how summary statistics are rounded. Values may be either integer(s)
or function(s). If not specified, default formatting is assigned
via \code{assign_summary_digits()}. See below for details.}

\item{...}{Arguments passed on to \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}}}

\item{language}{the language to use, defaults to Dutch}

\item{column1_name}{name to use for the first column}

\item{add_n}{add the overall N using \code{\link[gtsummary:add_n]{gtsummary::add_n()}}}

\item{add_p}{add the p values \code{\link[gtsummary:add_p]{gtsummary::add_p()}} (tests will be determined automatcally)}

\item{add_ci}{add the confidence interval using \code{\link[gtsummary:add_ci]{gtsummary::add_ci()}}}

\item{add_overall}{add the overall statistics using \code{\link[gtsummary:add_overall]{gtsummary::add_overall()}}}

\item{decimal.mark}{decimal separator, defaults to \code{\link[=dec_mark]{dec_mark()}}}

\item{big.mark}{thousands separator, defaults to \code{\link[=big_mark]{big_mark()}}}
}
\description{
Summarise a \link{data.frame} as \code{\link[gtsummary:tbl_summary]{gtsummary}} with Dutch defaults. These objects are based on the \code{gt} package by RStudio. To provide Certe style and compatibility with MS Word, use \code{\link[=tbl_flextable]{tbl_flextable()}} to transform the \code{\link[gtsummary:tbl_summary]{gtsummary}} object.
}
\details{
\code{\link[=tbl_gtsummary]{tbl_gtsummary()}} creates a summary table with \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}}, to which different extra columns can be added e.g. with \code{add_p = TRUE} and \code{add_overall = TRUE}.
}
\examples{
# These examples default to the Dutch language

iris |>
  tbl_gtsummary()

iris |> 
  tbl_gtsummary(Species, add_p = TRUE)
  
iris |> 
  tbl_gtsummary(Species, add_n = TRUE)
  
# support strata by providing 
iris2 <- iris
iris2$Category <- sample(LETTERS[1:2], size = 150, replace = TRUE)
head(iris2)

iris2 |> 
  tbl_gtsummary(c(Category, Species))

# transform to flextable 
# (formats to Certe style and allows rendering to Word)
iris |> 
  tbl_gtsummary(Species) |> 
  tbl_flextable()
}
