% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{tbl_markdown}
\alias{tbl_markdown}
\title{Print Table as Markdown, LaTeX of HTML}
\usage{
tbl_markdown(
  x,
  row.names = FALSE,
  column.names = colnames(x),
  align = NULL,
  padding = 2,
  caption = "",
  na = "",
  format.x = "markdown",
  format.dates = "dd-mm-yyyy",
  decimal.mark = ",",
  big.mark = ".",
  logicals = c("X", ""),
  columns.percent = NA,
  column.names.bold = TRUE,
  round.numbers = 2,
  round.percent = 1,
  newlines.leading = 0,
  newlines.trailing = 2
)
}
\arguments{
\item{x}{a \link{data.frame}}

\item{row.names}{row names to be displayed}

\item{column.names}{column names to be displayed}

\item{align}{alignment of columns (default: numbers to the right, others to the left)}

\item{padding}{extra cell padding}

\item{caption}{caption of table}

\item{na}{text for missing values (default: \code{""})}

\item{format.dates}{formatting of dates, will be evaluated with \code{\link[certestyle:format2]{certestyle::format2()}}}

\item{decimal.mark}{decimal separator}

\item{big.mark}{thousands separator}

\item{logicals}{vector with two values that replace \code{TRUE} and \code{FALSE}}

\item{columns.percent}{display the column indices as percentages using \code{\link[certestyle:format2]{certestyle::format2()}} - example: \code{columns.percent = c(2, 3)}}

\item{column.names.bold}{display column names in bold}

\item{round.numbers}{number of decimal places to round up for numbers}

\item{round.percent}{number of decimal places to round to when using \code{columns.percent}}

\item{newlines.leading}{number of white lines to print before the table}

\item{newlines.trailing}{number of white lines to print after the table}

\item{type}{type of formatting the table - valid options are \code{"latex"}, \code{"html"}, \code{"markdown"}, \code{"pandoc"} and \code{"rst"}}
}
\value{
character
}
\description{
Prints a \link{data.frame} as Markdown, LaTeX or HTML using \code{\link[knitr:kable]{knitr::kable()}}, with bold headers and Dutch number formats.
}
\details{
When in an R Markdown rapport a table is printed using this function, column headers only print well if \code{newlines.leading} >= 2, or by manually using \code{cat("\\\\n\\\\n")} before printing the table.
}
\seealso{
\code{\link[knitr:kable]{knitr::kable()}}
}
