% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update.data.frame}
\title{Update Data Set Based on Row Numbers}
\usage{
\method{update}{data.frame}(object, rows, ...)
}
\arguments{
\item{object}{a \link{data.frame}}

\item{rows}{row numbers or a \link{logical} vector}

\item{...}{arguments passed on to \code{\link[dplyr:mutate]{dplyr::mutate()}}}
}
\description{
Update a \link{data.frame} using specific integers for row numbers or a vectorised filter. Also supports dplyr groups. see \emph{Examples}.
}
\examples{
iris |> 
  update(3:4, Species = c("A", "B")) |> 
  head()
  
iris |> 
  update(Species == "setosa" & Sepal.Length > 5,
         Species = "something else") |> 
  head()
  
if (require("dplyr")) {

  # also supports dplyr groups:
  iris |> 
    group_by(Species) |>
    # update every 2nd to 4th row in group
    update(2:4, Species = "test") |> 
    # groups will be updated automatically
    count()

}
}
