% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{wikipedia_pageviews}
\alias{wikipedia_pageviews}
\title{Wikipedia Daily Page Views}
\usage{
wikipedia_pageviews(
  articles,
  date_range = last_6_months(),
  language = "nl",
  platform = c("all-access", "mobile-web", "mobile-app", "desktop"),
  agent = c("all-agents", "user", "spider", "automated")
)
}
\arguments{
\item{articles}{Wikipedia article title(s)}

\item{date_range}{defaults to \code{\link[=last_6_months]{last_6_months()}}}

\item{language}{defaults to \code{"nl"} for the Dutch Wikipedia}

\item{platform}{defaults to \code{"all-access"}. More reliable for epidemic trends might be \code{"mobile-web"}.}

\item{agent}{defaults to \code{"all-agents"}}
}
\description{
Download Wikipedia daily page views for any language, using the \href{https://doc.wikimedia.org/generated-data-platform/aqs/analytics-api/examples/page-metrics.html#page-views}{Wikimedia REST API}.
}
\details{
Zie voor een voorbeeld: \url{https://pageviews.wmcloud.org/?project=nl.wikipedia.org&platform=all-access&agent=all-agents&redirects=0&start=2024-06-01&end=2025-02-01&pages=Griep|Influenzavirus_A}
}
\examples{
wikipedia_pageviews("Griep")

articles <- c("Griep", "Koorts", "Hoest", "Verkoudheid", "Influenzavirus A",
              "Respiratoir syncytieel virus", "Keelpijn", "SARS-CoV-2")

if (require("certeplot2") && require("certestats") && require("dplyr")) {
  articles |>
    wikipedia_pageviews(c("2024-07-01", "2025-02-01")) |>
    group_by(article) |>
    mutate(z = z_score(views),
           z_ma = moving_average(z, w = 7, side = "left")) |>
    plot2(
      x = date,
      y = z_ma,
      category = article,
      type = "line",
      category.sort = "freq-asc",
      legend.position = "right",
      y.title = "7-daagse Z-score",
      x.title = "",
      title = "Wikipedia bezochte pagina's",
      subtitle = "Z-scores respiratoire seizoen")
}

if (require("certeplot2")) {
  wikipedia_pageviews("Dog",
                      language = "en",
                      date_range = last_5_years()) |>
    plot2(x = date, y = views)
}
}
