% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{word2pdf}
\alias{word2pdf}
\title{Convert Word-document to PDF}
\usage{
word2pdf(file, output_dir = NULL, overwrite = FALSE, teams_notice = FALSE)
}
\arguments{
\item{file}{file to be converted}

\item{output_dir}{target directory for PDF-file. Default is directory of Word-file.}

\item{overwrite}{overwrite existing PDF-file.}

\item{teams_notice}{send an update to this teams channel when conversion is complete. Use \code{NULL}, \code{NA} or \code{FALSE} to not send an update.}
}
\description{
Convert Word-document to PDF
}
\details{
this function returns the location of the PDF-file when conversion was succesful.
}
