% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{concat}
\alias{concat}
\title{Paste Elements Together}
\usage{
concat(...)
}
\arguments{
\item{...}{elements to be pasted together}
}
\description{
This function is identical to \code{paste(..., sep = "", collapse = "")}.
}
\examples{
concat("a", "b", "c")
}
