% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal.R
\name{concat}
\alias{concat}
\title{Paste Character Vectors Together}
\usage{
concat(...)
}
\arguments{
\item{...}{elements to be pasted together, can also be vectors}
}
\description{
This function is identical to \code{paste(c(...), sep = "", collapse = "")}.
}
\examples{
concat("a", "b", "c")

concat(c("a", "b"), "c")
}
