% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{import}
\alias{import}
\alias{import_rds}
\alias{import_excel}
\alias{import_xlsx}
\alias{import_csv}
\alias{import_csv2}
\alias{import_tsv}
\alias{import_spss}
\alias{import_clipboard}
\title{Import Data and Plots}
\usage{
import(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = FALSE,
  ...
)

import_rds(filename, card_number = project_get_current_id(ask = FALSE), ...)

import_excel(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  sheet = 1,
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import_xlsx(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  sheet = 1,
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import_csv(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import_csv2(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import_tsv(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import_spss(
  filename,
  card_number = project_get_current_id(ask = FALSE),
  auto_transform = FALSE,
  datenames = "en",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ".",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)

import_clipboard(
  sep = "\\t",
  header = TRUE,
  startrow = 1,
  auto_transform = TRUE,
  datenames = "nl",
  dateformat = "yyyy-mm-dd",
  timeformat = "HH:MM",
  decimal.mark = ",",
  big.mark = "",
  timezone = "UTC",
  na = c("", "NULL", "NA", "<NA>"),
  ...
)
}
\arguments{
\item{filename}{the full path of the file to be imported, will be parsed to a \link{character}}

\item{card_number}{a Trello card number}

\item{auto_transform}{transform the imported data with \code{\link[=auto_transform]{auto_transform()}}}

\item{...}{arguments passed on to methods}

\item{sheet}{Excel sheet to import, defaults to first sheet}

\item{datenames}{language of the date names, such as weekdays and months}

\item{dateformat}{expected date format, will be coerced with \code{\link[cleaner:format_datetime]{cleaner::format_datetime()}}}

\item{timeformat}{expected time format, will be coerced with \code{\link[cleaner:format_datetime]{cleaner::format_datetime()}}}

\item{decimal.mark}{separator for decimal numbers}

\item{big.mark}{separator for thousands}

\item{timezone}{expected time zone}

\item{na}{values to interpret as \code{NA}}

\item{sep}{character to separate values in a row}

\item{header}{use first row as header}

\item{startrow}{first row to start importing}
}
\description{
These functions can be used to import data. They work closely with the \code{certeprojects} package and support Trello card numbers.
}
\details{
The \code{\link[=import]{import()}} function uses the \code{rio} package to guess the required import function.

The \code{\link[=import_clipboard]{import_clipboard()}} function requires the \code{clipr} package to be installed.
}
\examples{
#
}
