% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_secret}
\alias{read_secret}
\title{Read Certe Secret From File}
\usage{
read_secret(property)
}
\arguments{
\item{property}{the property to read, case-sensitive}
}
\description{
This function reads from a local or remote YAML file, as set in the environmental variable \code{"secrets_file"}.
}
\details{
This function will return a vector of length > 1 if the property value contains comma's. The property name and value have to be separated with a colon (\code{:}).
}
\examples{
# for this example, a temporary 'secrets' file
my_secrets_file <- tempfile(fileext = ".yaml")
Sys.setenv(secrets_file = my_secrets_file)
writeLines("tenant_id: 8fb3c03060e02e89", my_secrets_file)
writeLines("default_users: user_1, user_2", my_secrets_file)

read_secret("tenant_id")
read_secret("default_users")
}
