% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal.R
\name{read_secret}
\alias{read_secret}
\title{Read Certe Secret From File}
\usage{
read_secret(property, file = Sys.getenv("secrets_file"))
}
\arguments{
\item{property}{the property to read, case-sensitive}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing}
}
\description{
This function reads from a local or remote YAML file, as set in the environmental variable \code{"secrets_file"}.
}
\details{
In the secrets file, the property name and value have to be separated with a colon (\code{:}), as is intended in YAML files.

The default value for \code{file} is the environmental variable \code{"secrets_file"}.

The file will be read using \code{\link[yaml:read_yaml]{read_yaml()}}, which allows almost any local path or remote connection (such as websites).
}
\examples{
# for this example, create a temporary 'secrets' file
my_secrets_file <- tempfile(fileext = ".yaml")
Sys.setenv(secrets_file = my_secrets_file)
writeLines(c("tenant_id: 8fb3c03060e02e89",
             "default_users: user_1"),
           my_secrets_file)

read_secret("tenant_id")
read_secret("default_users")
}
