% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remember}
\alias{remember}
\alias{recall}
\title{Temporarily Remember Object}
\usage{
remember(.data, ...)

recall(x = NULL, delete = TRUE)
}
\arguments{
\item{.data}{\link{data.frame}}

\item{...}{value(s) to be remembered}

\item{x}{value to be recalled}

\item{delete}{a \link{logical} to indicate whether the delete value after recalling}
}
\description{
Can be used in dplyr-syntax to remember values and objects for later use. Objects are (temporarily) stored in the \code{certetoolbox} package environment.
}
\details{
values can be saved with \strong{\code{\link[=remember]{remember()}}} and recalled (and deleted) with \strong{\code{\link[=recall]{recall()}}}.
}
\examples{
\dontrun{
 tbl \%>\%
   filter(...) \%>\%
   remember(rows = nrow(.)) \%>\%
   group_by(...) \%>\%
   summarise(...) \%>\%
   plot2(title = "Test",
         subtitle = paste("n =", recall(rows)))
}
}
