% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{remember}
\alias{remember}
\alias{recall}
\title{Temporarily Remember Objects}
\usage{
remember(.data, ...)

recall(x = NULL, delete = TRUE)
}
\arguments{
\item{.data}{\link{data.frame}}

\item{...}{value(s) to be remembered}

\item{x}{value to be recalled}

\item{delete}{a \link{logical} to indicate whether the delete value after recalling}
}
\description{
Can be used in dplyr-syntax to remember values and objects for later use. Objects are (temporarily) stored in the \code{certetoolbox} package environment.
}
\details{
values can be saved with \strong{\code{\link[=remember]{remember()}}} and recalled (and deleted) with \strong{\code{\link[=recall]{recall()}}}.
}
\examples{
library(dplyr, warn-conflicts = FALSE)

x <- mtcars \%>\% remember(nrow(.)) 
recall()
recall() # value removed
x <- mtcars \%>\% remember(n = nrow(.)) 
recall(n)
recall(n) # value removed

\dontrun{
 tbl \%>\%
   filter(...) \%>\%
   remember(rows = nrow(.)) \%>\%
   group_by(...) \%>\%
   summarise(...) \%>\%
   plot2(title = "Test",
         subtitle = paste("n =", recall(rows)))
}
}
