% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{size_humanreadable}
\alias{size_humanreadable}
\title{Human-readable File Size}
\usage{
size_humanreadable(bytes, decimals = 1, decimal.mark = ",")
}
\arguments{
\item{bytes}{number of bytes}

\item{decimals}{precision, not used for bytes and kilobytes}

\item{decimal.mark}{decimal mark to use}
}
\description{
Formats bytes into human-readable units, from "kB" (10^3) to "YB" (10^23).
}
\examples{
size_humanreadable(c(12, 1234, 123456, 12345678))

size_humanreadable(1024 ^ c(0:4))
}
