% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{tbl_gtsummary}
\alias{tbl_gtsummary}
\title{Summarise Table as \code{gtsummary}}
\usage{
tbl_gtsummary(
  x,
  by = NULL,
  label = NULL,
  digits = 1,
  ...,
  language = "nl",
  column1_name = "Eigenschap",
  add_n = FALSE,
  add_p = FALSE,
  add_ci = FALSE,
  add_overall = FALSE,
  decimal.mark = dec_mark(),
  big.mark = big_mark()
)
}
\arguments{
\item{x}{a \link{data.frame}}

\item{by}{A column name (quoted or unquoted) in \code{data}.
Summary statistics will be calculated separately for each level of the \code{by}
variable (e.g. \code{by = trt}). If \code{NULL}, summary statistics
are calculated using all observations. To stratify a table by two or more
variables, use \code{tbl_strata()}}

\item{label}{List of formulas specifying variables labels,
e.g. \code{list(age ~ "Age", stage ~ "Path T Stage")}.  If a
variable's label is not specified here, the label attribute
(\code{attr(data$age, "label")}) is used.  If
attribute label is \code{NULL}, the variable name will be used.}

\item{digits}{List of formulas specifying the number of decimal
places to round summary statistics. If not specified,
\code{tbl_summary} guesses an appropriate number of decimals to round statistics.
When multiple statistics are displayed for a single variable, supply a vector
rather than an integer.  For example, if the
statistic being calculated is \code{"{mean} ({sd})"} and you want the mean rounded
to 1 decimal place, and the SD to 2 use \code{digits = list(age ~ c(1, 2))}. User
may also pass a styling function: \code{digits = age ~ style_sigfig}}

\item{...}{Arguments passed on to \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}}}

\item{language}{the language to use, defaults to Dutch}

\item{column1_name}{name to use for the first column}

\item{add_n}{add the overall N using \code{\link[gtsummary:add_n]{gtsummary::add_n()}}}

\item{add_p}{add the p values \code{\link[gtsummary:add_p]{gtsummary::add_p()}} (tests will be determined automatcally)}

\item{add_ci}{add the confidence interval using \code{\link[gtsummary:add_ci]{gtsummary::add_ci()}}}

\item{add_overall}{add the overall statistics using \code{\link[gtsummary:add_overall]{gtsummary::add_overall()}}}

\item{decimal.mark}{decimal separator, defaults to \code{\link[=dec_mark]{dec_mark()}}}

\item{big.mark}{thousands separator, defaults to \code{\link[=big_mark]{big_mark()}}}
}
\description{
Summarise a \link{data.frame} as \code{\link[gtsummary:tbl_summary]{gtsummary}} with Dutch defaults. These objects are based on the \code{gt} package by RStudio. To provide Certe style and compatibility with MS Word, use \code{\link[=tbl_flextable]{tbl_flextable()}} to transform the \code{\link[gtsummary:tbl_summary]{gtsummary}} object.
}
\details{
\code{\link[=tbl_gtsummary]{tbl_gtsummary()}} creates a summary table with \code{\link[gtsummary:tbl_summary]{gtsummary::tbl_summary()}}, to which different extra columns can be added e.g. with \code{add_p = TRUE} and \code{add_overall = TRUE}.
}
\examples{
# These examples default to the Dutch language

iris |>
  tbl_gtsummary()

iris |> 
  tbl_gtsummary(Species, add_p = TRUE)
  
iris |> 
  tbl_gtsummary(Species, add_n = TRUE)
  
# support strata by providing 
iris2 <- iris
iris2$Category <- sample(LETTERS[1:2], size = 150, replace = TRUE)
head(iris2)

iris2 |> 
  tbl_gtsummary(c(Category, Species))

# transform to flextable 
# (formats to Certe style and allows rendering to Word)
iris |> 
  tbl_gtsummary(Species) |> 
  tbl_flextable()
}
